refactor(dts-tests): split constructor object type tests
