+++ crates/tsz-emitter/src/declaration_emitter/tests/mod.rs
@@ -304,3 +304,4 @@ mod probes_tsc_comparison;
 mod simple_declarations;
 mod type_formatting;
 mod type_info;
+mod type_info_constructor_objects;
